fix rename dir
