+++ lib/ui/error_handler_mixin.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:smart_order_app/domain/errors/error.dart';
-import 'package:smart_order_app/ui/component/snackBar/simple_snackbar.dart';
+import 'package:smart_order_app/ui/component/snack_bar/simple_snackbar.dart';
 
 mixin ErrorHandlerMixin {
   Future<void> execute(
